fix: one more 'unsigned docid' issue
That is consequence of 'interpret int as uint' - that is broken when
comparing numbers (i.e. now we should always cast int to uint when
comparing, that is non-obvious and error-prone. Until we convert DocID
to something really 'unsigned', such kind of defects will rise again and
 again.

Related to #3181 (the issue is actually fixed, but boring checklist
still need to be completed)