fix!: rename mva to multi
Historically multi-value attributes were
displayed as 'mva' or 'mva64'. However, everywhere else they were
'multi' and 'multi64'; especially when creating a table in RT mode.
Since change is mostly cosmetic, it might not be critical, but if some
soft depends on that naming, it will be breaking change for it.

ref #2281